PYTHON/REMOTE-PC-WEB-STATS-LINUX - docker (NUC-I3)/requirements.txt
flask
psutil
gevent
gunicorn
dbus-python
flask-cors
